requirements.txt
bcrypt==4.0.1
beautifulsoup4==4.11.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
commonmark==0.9.1
cryptography==39.0.0
Deprecated==1.2.13
fandom-py==0.2.1
Flask==2.2.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
limits==3.2.0
MarkupSafe==2.1.2
memory-profiler==0.61.0
ordered-set==4.1.0
packaging==23.0
paramiko==3.0.0
psutil==5.9.4
pycparser==2.21
Pygments==2.14.0
pymediawiki==0.7.2
PyNaCl==1.5.0
pysftp==0.2.9
python-dotenv==0.21.1
requests==2.28.2
rich==12.6.0
soupsieve==2.3.2.post1
typing_extensions==4.4.0
urllib3==1.26.14
waitress==2.1.2
Werkzeug==2.2.2
wrapt==1.14.1
